+++ src/methods/edit_message_text.rs
@@ -13,7 +13,7 @@ pub struct EditMessageText<'a> {
     #[serde(skip)]
     proxy: Option<proxy::Proxy>,
     chat_id: types::ChatId<'a>,
-    message_id: u64,
+    message_id: u32,
     text: &'a str,
     #[serde(skip_serializing_if = "Option::is_none")]
     parse_mode: Option<types::ParseMode>,
@@ -28,7 +28,7 @@ impl<'a> EditMessageText<'a> {
     pub fn new(
         token: &'a str,
         chat_id: impl Into<types::ChatId<'a>>,
-        message_id: u64,
+        message_id: u32,
         text: &'a str,
     ) -> Self {
         Self {
