+++ src/types/audio.rs
@@ -8,7 +8,7 @@ pub struct Audio {
     /// The file ID of the audio.
     pub file_id: String,
     /// The duration of the audio.
-    pub duration: i64,
+    pub duration: u32,
     /// The performer of the audio.
     pub performer: Option<String>,
     /// The title of the audio.
@@ -16,7 +16,7 @@ pub struct Audio {
     /// The MIME type of the audio.
     pub mime_type: Option<String>,
     /// The audio file's size.
-    pub file_size: Option<i64>,
+    pub file_size: Option<u32>,
     /// The thumb of the audio.
     pub thumb: Option<PhotoSize>,
 }
