+++ src/types/sticker.rs
@@ -8,9 +8,9 @@ pub struct Sticker {
     /// The file ID of the sticker.
     pub file_id: String,
     /// The width of the sticker.
-    pub width: i64,
+    pub width: u32,
     /// The height of the sticker.
-    pub height: i64,
+    pub height: u32,
     /// The thumb of the sticker.
     pub thumb: Option<PhotoSize>,
     /// The emoji of the sticker.
@@ -20,5 +20,5 @@ pub struct Sticker {
     /// The sticket's position if it is a mask.
     pub mask_position: Option<MaskPosition>,
     /// The sticker file's size.
-    pub file_size: Option<i64>,
+    pub file_size: Option<u32>,
 }
